Convert distance units from cm to meters
Since we generally want to use base SI units, this converts all distance
measurement units from centimeters to meters.

+++ src/hal/lidar_scanner.hpp
@@ -19,7 +19,7 @@ public:
         double distance;
         double angle;
 
-        SamplePoint(uint16_t distance, double angle) : distance(distance), angle(angle) {}
+        SamplePoint(double distance, double angle) : distance(distance), angle(angle) {}
     };
 
     using Scan = std::shared_ptr<std::vector<SamplePoint>>;
